# apis/beans/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Beans is a publisher-content API for news, blogs, financial and earnings reports, litigation and lawsuits, official statements, research, technical documents, and related coverage context.\nBeans finds and verifies what publishers published. It returns citable Articles, Source metadata, attention-ranked feeds, similar publisher reading, external Article mentions, and normalized filter discovery.\nCollections return `{data, pagination, meta}`. pagination contains `limit`, `num_results` (this page only), and `next_cursor`. Empty collections return HTTP 200 with `data: []`. Missing detail resources return HTTP 404. Errors return `{ \"error\": { \"code\", \"message\" } }`.\n`content_type=post` is not a valid request filter. `post` may still appear on Article responses. Unknown or route-inapplicable query parameters return HTTP 400.\nBackend authentication uses the `X-API-KEY` header (or other headers listed in `API_KEY`). `/health` does not require a key. Public clients send Bearer keys to the gateway, not this service.",
        "title": "Beans News API \u0026 MCP",
        "contact": {
            "name": "Project Cafecito",
            "url": "https://cafecito.tech",
            "email": "soumitsrah@cafecito.tech"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/sources/{id}": {
            "get": {
                "security": [
                    {
                        "BackendAPIKey": []
                    }
                ],
                "description": "Returns one publisher Source selected by UUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sources"
                ],
                "summary": "Get a Source",
                "operationId": "getSource",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Source UUID.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/router.SourceDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Source not found",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Service unavailable",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "router.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "invalid_request"
                },
                "message": {
                    "type": "string",
                    "example": "Unknown or unsupported query parameter: offset"
                }
            }
        },
        "router.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/router.ErrorBody"
                }
            }
        },
...
        "router.SourceDetailResponse": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/router.SourceDocument"
                }
            }
        },
        "router.SourceDocument": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "favicon_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "name": {
                    "type": "string"
                },
                "rss_feed_url": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BackendAPIKey": {
            "type": "apiKey",
            "name": "X-API-KEY",
            "in": "header"
        }
    }
}